chore(devcontainer): move Dockerfile under .devcontainer and fix compose/devcontainer.json paths
