+++ config/next-sitemap.config.cjs
